move text gen files to generate
